+++ tests/HelperEnvTrait/HelperEnvTestingTest.php
@@ -18,11 +18,11 @@ final class HelperEnvTestingTest extends TestCase
     #[Test]
     public function envTesting(): void
     {
-        Helper::cacheRuntimeSet(HelperInternal::class . 'internalEnv', ['APP_ENV' => 'Testing']);
+        Helper::cacheRuntimeSet(HelperInternal::class . 'internalEnv' . Helper::hashXxh128(''), ['APP_ENV' => 'Testing']);
         $boolean = Helper::envTesting();
         $this->assertTrue($boolean === true);
 
-        Helper::cacheRuntimeSet(HelperInternal::class . 'internalEnv', ['APP_ENV' => 'none']);
+        Helper::cacheRuntimeSet(HelperInternal::class . 'internalEnv' . Helper::hashXxh128(''), ['APP_ENV' => 'none']);
         $boolean = Helper::envTesting();
         $this->assertTrue($boolean === false);
     }