+++ CHANGELOG.md
@@ -1,10 +1,14 @@
 # Changelog
 
+## _Версия 1.19_
+- Обновлен метод Helper::timeSecondsToString
+
 ## _Версия 1.18_
 - Добавлен метод Helper::arrayDeleteValues
 - Добавлен метод Helper::arrayFlip
 - Добавлен метод Helper::stringSplitRange
 - Добавлен метод Helper::timeFromSeconds
+
 - Переименован метод Helper::numberSwap в Helper::varSwap
 
 ## _Версия 1.17_